function_app/requirements.txt
azure-functions==1.18.0
azure-cosmos==4.5.1
azure-identity==1.15.0
requests==2.31.0 